+++ src/universal/Http/ClientInterface.php
@@ -25,7 +25,13 @@ use Psr\Http\Message\ResponseInterface;
 
 /**
  * ClientInterface is a contract to create object representing the client in the server side. The client must be
- * agnostic and accepts only \Throwable exception and PSR7 response.
+ * agnostic and accepts only \Throwable exception and PSR7 response. It's possible to pass a PSR7 Response object
+ * without send it via the method "acceptResponse".
+ *
+ * To update an response, it's mandatory to call the method
+ * "updateResponse" and pass a callable able to update the response and update it into the client.
+ *
+ * The method "sendResponse" as a behavior like updateResponse but send directly the response.
  *
  * All public method of the client must only return the self client or a clone instance.
  *
@@ -39,13 +45,18 @@ use Psr\Http\Message\ResponseInterface;
 interface ClientInterface
 {
     /**
+     * To update a response in the client. You must pass a callable able to work on the response passed into argument,
+     * and repush it into the client via its method "acceptResponse". (Response object are immutable).
+     * The callable must accept a ClientInterface object as first argument, and a PSR7 Response as second (can be null).
+     *
      * @param callable $modifier
+     *
      * @return ClientInterface
      */
     public function updateResponse(callable $modifier): ClientInterface;
 
     /**
-     * To accept a response from the controller action and send it to the HTTP client.
+     * To accept a response from the controller action without send it to the HTTP client.
      *
      * @param ResponseInterface $response
      *
@@ -56,11 +67,12 @@ interface ClientInterface
     /**
      * To accept a response from the controller action and send it to the HTTP client.
      *
-     * @param ResponseInterface $response
+     * @param ResponseInterface|null $response
      *
      * @return ClientInterface
+     * @throws \RuntimeException when no response was been defined via acceptResponse and $response argument is null.
      */
-    public function sendResponse(ResponseInterface $response): ClientInterface;
+    public function sendResponse(ResponseInterface $response=null): ClientInterface;
 
     /**
      * To intercept an error during a request and forward the message to the HTTP client.